(PUP-8501) Remove uses of to_json that rely on built-in JSON gem
The built-in JSON gem monkey patches many of Ruby's built-in types to
make them provide `to_json` methods. However, not all of the backends
that MultiJson can use do this. To avoid calling `to_json` where it
might not exist, this commit removes all cases where `to_json` might get
called on a built-in Ruby type. It leaves cases where we are calling a
custom `to_json` defined on one of our own types.