(PUP-8509) Add flatten() function from stdlib to puppet
This adds the flatten function from stdlib to puppet.
This version is a 4.x implementation of the function and it is backwards
compatible. It adds a new feature in that it will concatenate and
flatten multiple arguments, and this also turns non array values into
being wrapped into an array.