(PUP-8501) Avoid calling `to_hash` on JSON state objects
MultiJson calls `merge` on the options it is passed, which relies
on the options' defining a `to_hash` method. In Ruby 1.9.3,
JSON::Ext::Generator::State only defines `to_h`, not `to_hash`, so we
need to convert it first, similar to what is done when handling the `pretty`
option. Later versions of the JSON gem alias `to_h` to `to_hash`, so this
can be removed once we drop Ruby 1.9.3 support.