(maint) Add trace, clearer message to EOFError
This is an improvement for PUP-3238, though not a solve.

Net::HTTP will raise an EOFError when something goes wrong at the
network stack level, such as a connection being interrupted or
terminated by a load balancer in the middle of a transaction. By
default, Ruby raises an EOFError with the message "end of file reached"
and no backtrace.

This commit catches EOFError exceptions raised by Net::HTTP#request, and
re-raises the EOFError with a meaningful message and backtrace. Because
this type of error is typically caused by problems from the network
stack level, which are frequently timeout-related, the error message
will include how long the request ran before termination.

At the very least it will be much clearer what is happening when these
kinds of errors occur.

Previously:
Error: Could not retrieve catalog from remote server: end of file
  reached

Now:
Error: Could not retrieve catalog from remote server: request
  https://master.llu.puppet.vm:9140/puppet/v3/catalog/master.llu.puppet.vm
  interrupted after 1.004 seconds