(PUP-8537) Change match expression validation to have side effect
Before this, it was considered an error to have a match expression
followed by an expression that uses the numeric variables. This was
caused because the "idem" validation did not consider that match
expression have a potential side effect on the numeric match variables
(setting or clearing them).

This is not changed.

It is however not possible (with reasonable effort and cost at runtime)
to authoritatively find cases when a match does not modify the numeric
variables. It therefore allows all match expressions even if some could
have no side effect and thus be a potential error.