(PUP-8538) Remove chpasswd username for AIX
Prior to this commit, there was an invalid parameter in the `chpasswd`
command for the AIX user resource. This is a problem in newer versions
of AIX that have fixed the `chpasswd` to return a failure when there are
extraneous parameters in the command. This commit removes the invalid
parameter to allow for passwords to be set in the user resource on AIX.