(PUP-7675) Add support for convert_to lookup option
This adds support for convert_to in lookup_options as a lookup_option
key.

The lookup_options now has a convert_to key that can be 
either String (a data type string), or Tuple[String, Any, 1].

The result is only converted at the very end.
There is no conversion of a default value given to the lookup function,
or returned from a block computing the default value - they must return
the expected result type.