(PUP-7675) Refactor the handling of convert_to
This optimizes the handling of convert_to by moving the conversion logic
to be inside of the LookupAdapter class instead of in the Lookup helper
class. This is better because the lookup adapter must (for all APL
lookups) also lookup the lookup_options. By moving the logic, the
convert_to lookup can piggy-back on that lookup.