(PUP-8541) Make lookup of yaml_data fail on bad data when --strict=error
Prior to this commit, a warning would be logged when data that was read
by the lookup function `yaml_data` or `eyaml_lookup_key` wasn't in the
form of a hash. This commit retains that behavior with the addition that
if `--strict=error` is given, and the data doesn't indicate an empty
file (`YAML#load` returns `false` for this), then an error will be
raised.