(PUP-7675) Add test for nil value not being converted by convert_to
If nil values were to be converted it would be impossible to override
with a nil value using data. The opposite is however possible - a user
can add whatever result that was expected as the outcome of a nil to
some other data type - for example empty array, empty string.