(PUP-8501) Remove force_encoding when dumping to JSON
The MultiJson library does a `force_encoding` to ASCII-8BIT for all
strings that it is converting to JSON, due to a somewhat-arcane history
of tried and failed features. Originally we were replicating this in our
JSON wrapper, but on further inspection we were unable to find a good
reason to continue doing so. This commit removes the `force_encoding`.

Note that it will still occur when using MultiJson library instead of
falling back to our wrapper code.