(PUP-723) Check to see if a repo config file has already been created when creating a yumrepo resource
In the case that a yumrepo config file that we will be managing a section in is
created after prefetch has happened, any yumrepo resource that follows will be
operating from a stale cache (virtual_inifile). This will cause any content in
the file for any section that we are managing to be cobbered. This is because
when a yumrepo resource is flushed, a new file is written to disk based on what
is in the virtual_inifile instead of modifying the existing file on disk.  Now
we refresh the virtual_inifile if the target of the yumrepo exists on disk but
does not exist in the virual_inifile.