(PUP-8549) Force tag validation to be done in UTF-8
Previously, when checking if a string was a valid tag we would not
ensure the encoding of either the tag string or the validation regex.
Usually, when one is in a "compatible" encoding of the other, ie the
string is in ASCII_8BIT and the regex is in UTF_8 Ruby will attempt to
"do the right thing" and convert one to the other.

However, this implicit converstion may convert the regex to the value
of the string, which means a new regex will be compiled each time a
string with a different encoding is reached. And because we will often
read manifests of one encoding and then create strings out of them into
a different encoding we may end up recompiling the regex frequently
during a catalog compilation. This is primarily a problem because the
character class `[[:alnum:]]` that we use in tag validation is gigantic
in UTF-8.

To disallow, or at least limit, regex recompilation we want strings that
we compare to the validation regex to be encoded in UTF-8 whenever
possible. To do so we wrap the regex matching in a method and encode our
inputs, with our inputs forced to UTF-8 we also force the regex to UTF-8
for consistency.