(PUP-8476) Make EPP comment not trim left by default and add <%#-
This changes the default behaviour of an EPP comment `<%#` to not trim
left by default as that is both not compatible with ERB and makes it
hard to use when spaces should be kept.

To make it possible to trim left, a new EPP tag `<%#-` has been added
that will make the EPP comment behave as it did prior to this change.