(PUP-8514) Define sensitive parameters globally
Prior to this change, we could only define an instance of a parameter as
sensitive. This was a pain because we'd have to go back once parameter
instances had been created and specifically set them as sensitive. With
this change, we can now define a parameter as sensitive when creating
the class. With this change, all instances of a class marked as
sensitive will automatically be defined as sensitive! This will help us
reduce some loops we previously had to jump through in order to actually
mark a parameters as sensitive.