(maint) Mark user passwords as sensitive
With the change introduced in PUP-8514, we no longer have to
retroactively define an instance of the password as sensitive. All
instances are now marked sensitive since we can define this at the
class level now.

With this change, we can also remove the custom `is_to_s`, `should_to_s`,
and `change_to_s` methods. They had previously been used to redact any
password information in the logs. With this change, we're correctly
marking passwords as sensitive information, and we can take advantage of
existing logic to redact passwords.