(PUP-8588) Support multiple implementations for a single task
Previously, tasks could only have a single executable and a metadata
file. Any task with more than one executable would fail to load. In
order to support cross-platform tasks, we now allow tasks to have
multiple implementations, differentiated with the new 'implementations'
key in the metadata.

The 'implementations' key is a list of executable files, along with a
set of 'requirements', which are implementation-defined features that
necessary to run that version of the task. A task can have multiple
executables, so long as the implementations key is present and lists
them. If there are multiple executables but no implementations key, the
task will fail to load as before.

Any task executables that *aren't* mentioned in the implementations list
are ignored. This helps avoid the scenario where swap files, etc could
cause a task to fail to load. It is an error for the implementations
list to specify an executable file that doesn't exist.

A task with a single executable may have an implementations list, but
doesn't need one. If not specified, the single executable will be
used as the only available implementation, and will be considered
suitable on all targets.