(PUP-8561) Augeas provider: unescape escaped quotes
The argument for commands like 'set' can be enclosed in single or double
quotes. To place the quote character into the string, they can be escaped
with a backslash.

So far, the provider did not remove the backslash-escape from the string,
making it impossible for users to have, e.g., a plain double quote inside a
double-quoted argument.

With this change, backslash-escaped quote characters are unescaped from
strings that are enclosed with that quote character; e.g., a
backslash-escaped double quote is reduced to just a double quote in a
string enclosed in double quotes (but left untouched if the string
is enclosed in single quotes)

Fixes https://tickets.puppetlabs.com/browse/PUP-8561