(PUP-8486) Catch undefined_variable in hiera's scope
Before this, the general handling of undefined variable caused the throw
of :undefined_variable to be translated into an exception  - which in
turn was handled under the control of Puppet[:strict], leading to a
warning or error (or no action).

When running on j9k JRuby it has been showed that the exception is aprox
70 times slower than catching the symbol, and thus when running with
strict mode off or warning, the performance suffers greatly.

This commit adds a catch around the call to `lookupvar` in the special
scope used with hiera and issues the same warning/error under control of
Puppet[:strict] as the original implementation in the real Scope.