(PUP-8562) Do not overpopulate context stack
Prior to this commit, we were retrieving and storing the puppet server
version for every request made. Unfortunately, we were never popping any
of this data off the stack. This resulted in a hugely overpopulated
stack that caused a significant slowdown in catalog application time.

This commit ensures that we only add the server version to the stack
once. It's available to reference later on, and we don't add the extra
overhead of trying to re-add it every time.

For additional context, the agent is retrieving the server's Puppet
version from the response header, and permanently adding that to the
context so it can be looked up later to prevent requests the server
can't handle. We don't support different server versions handling separate
requests, so we only need to add this once to make it available for the
rest of the agent run.