(maint) Drop strict=off around tests no longer needing this
This removes strict=off in a before around relationship tests
that always error. This is done since the behavior is no longer
under control of the strict flag and the change from being under control
of strict happened several releases ago.

It is just confusing to test with strict=off for behavior that is the
same for all settings of strict (we are not testing this for all other
things in the code base).