(PUP-8610) Speedup PuppetStack by adding top_of_stack and use that
All place where the stacktrace is used the logic is only interested in
getting the file, line from the top of the stack.

This commit adds a top_of_stack method to PuppetStack and makes use of
that in all locations where file & line are needed.

If it proves that getting the entire caller() stack is time consuming,
it can be changed to loop over the history in batches of n.