(TBD) Prepare for modules to be moved to vendor directory
Adds a load_modules.rb shim that appends the lib directory for all
vendored modules to the load path. The shim is automatically required by
vendor.rb, so that vendored modules are always on the $LOAD_PATH.

Updates the rake spec and parallel:spec tasks to include vendored
modules specs if an explicit list of tests to run is not given.