(PUP-8637) Allow merge of definitions to main when --freeze_main is on
Before this, in some circumstances puppet ends up wanting to merge
definitions into main. When freeze_main is on, the check for allowed
merge was very simplistic and did not allow for definitions to be
merged. This is always ok since they are just definitions that does not
lead to any declaration of resources in the catalog.

One test is modified as it relied on the very simple freeze_main
conditions. One test was added to test the allowed (definitions only)
conditions that should allow merge to take place.