(PUP-8614) Remove 'should return a new X509 certificate' test
This test appeared to just be testing if a new object would be returned from
the .build command of Puppet::SSL::CertificateFactory.

Now that the OpenSSL certificate objects have real object comparison this test
started failing because we are passing in the same csr/serial (which would
create the same certificate) and the comparison was matching as true.

Since the nature of the test doesn't seem to have much use it's been removed