(PUP-8653) Use new http client to download certificates
This commit updates SSL::Host to use the new rest client to download
certificates from the CA, instead of going through the indirector.

The logic in this commit is more complicated than it will ultimately
need to be, because at this point we still have a Ruby CA
implementation, and its associated tests use indirector termini other
than file and rest, which are harder to replace. The special casing
around `ca_location` and `terminus_class` can be removed along with the
Ruby CA.