(PUP-6344) Always populate total time metric in reports
Previously if anything raised during a run, the total time would be nil,
as the thinmark wouldn't be able to return & set run_internal_time. To get
around this we're now manually recording when the run_internal starts, and
comparing that against Time.now at the end to record the total run time.