(maint) Clarify error on Timeout::Error exception
This is another improvement for PUP-3238, though not a solve.

It turns out Net::HTTP will raise other errors, such as Timeout::Error, in
addition to the previously caught EOFErrors. To further clarify to the user
what is happening when these network issues occur, this commit also catches and
adds contexual information to Timeout::Error exceptions.

For completeness, in the event any other type of exception is raised we also
catch it and add the uri context. Seconds elapsed is omitted in the general
case because not knowing what the exception is, it's not certain whether or
not seconds elapsed will be relevant.