(PUP-6344) Ensure the total time is added to the report
Previously, if we exited the `run_internal` method before reaching the
`apply_catalog` method, then the report was never finalized. This could
happen if we failed to retrieve a catalog or if the prerun command
failed.

This commit ensures we add a total time metric based on wall clock time
between when the run_internal method starts and when we finalize the
report, just prior to sending it.

This commit also preserves the behavior whereby Configurer#run returns
nil if we didn't apply the catalog, for example, if there is a cycle in
the catalog. Otherwise it returns whatever Report#exit_status returns,
which is either -1, 0, or a bitwise combination of 2 and 4.