(PUP-8582) Add vendor_modules to basemodulepath
Append the vendor_modules directory to the basemodulepath setting so
that vendored modules are available in all environments.

On most platforms, the resulting basemodulepath corresponds to global,
site and vendored modules directories, respectively.

On Windows, the resulting basemodulepath contains at least the global
modules directory. Additionally if the env_windows_installdir fact[1] is
set, which it is when running from puppet-agent packages, then we
include the site and vendored modules directories.

For all platforms, the global modules directory ($codedir/modules)
depends on whether we are running in a privileged context or not (so
it's either in /etc/puppetlabs or ~/.puppetlabs).

[1] https://puppet.com/docs/facter/3.11/core_facts.html#envwindowsinstalldir