(PUP-8664) rpm provider uninstall too specific
Prior to this commit:

The rpm provider used potentially out-of-date cache information when uninstalling.
The uninstall command options were more specific than the resource for version and release.
The if statement for arch always appended arch to the uninstall command options.

With this commit:

The cache is updated before use in the uninstall method.
The install is as specific as the resource for version and release and arch.
The spec test correctly tests both specifying and not specifying arch.