(PUP-1952) Update dpkg provider spec for jRuby
jRuby doesn't actually call new inside of Tempfile.open. Instead it asks
Java for a new instance of the class, and calls initialize on it. To stub
out what Tempfile is doing in both MRI & jRuby, we need to stub out
Tempfile.open itself, and have it yield our mock object.