(PUP-8558) Remove strict_semver as an option from PMT and always enforce strict semver logic.
This commit removes the option to have relaxed semver option in the puppet
module tool. When this was added it caused problems because puppet agent,
puppet server, and puppet all store and look for semantic_puppet in different
locations and with different methods with differing number of arguments, which
caused major discrepancies between versions. This also caused some semver
issues with installing directly from the forge with the PMT. This is the first
step to completely unvendoring semantic_puppet so that it can exist in only one
place and all three products get it from the same location. Removed all
references to strict_semver as well as related logic around pre-release
versioning, code checking to see if this option was set, as well as all related
tests for this option.