(PUP-8663) Disable ruby's http_keepalive_timeout
Ruby 2 added a keep_alive_timeout accessor to Net::HTTP which defaults
to 2 seconds. If the span between when the connection was last used and
when we try to use it again is greater than the timeout, then ruby will
automagically close and reopen a new TCP/SSL connection. This interferes
with puppet's http pool of persistent connections, and leads to
unnecessary load on the puppetserver.

This scenario is common if fact evaluation takes more than 2 seconds,
causing a new connection to be made for the catalog request.

This commit disables ruby's http_keepalive_timeout so that we can always
manage the lifetime of the TCP/SSL connection. The constant 2**31 - 1
was chosen as the maximum 32-bit signed value, which is big enough for
our needs, works on 32-bit architectures, and can be compared against
Time.now in Ruby 2.1[1] and Process.clock_gettime(Process::CLOCK_MONOTONIC)
in Ruby 2.4[2].

[1] https://github.com/ruby/ruby/blob/v2_1_9/lib/net/http.rb#L1454
[2] https://github.com/ruby/ruby/blob/v2_4_3/lib/net/http.rb#L1487