(PUP-8688) Add utility for selecting server and port
This commit adds a ServerResolver class responsible for finding the
correct server and port to use to query a given API, represented by a
Route object. Client instances must be created with a particular route,
which entails that a different HTTP client should be used for making
requests to different APIs, e.g. the CA server versus the Puppet
master.