(PUP-8685) Make .gemspec buildable
Previously, the .gemspec was only there so that Puppet could
work with Bundler workflows (i.e. be specified as a dependency
in a Gemfile that points to a git ref.). It was not meant to be
buildable. Instead, puppetlabs/packaging's pl:gem task was used
to build the shipped Puppet gem so that all of the gem information
was specified in ext/project_data.yaml.

This was unnecessarily complicated and extremely odd -- one would
expect that a 'gem build .gemspec' would build the gem. This PR removes
all of the gem stuff out of ext/project_data.yaml into the .gemspec so
that it is buildable with `gem build .gemspec`. The only gem-specific information
in ext/project_data.yaml are the gem_platforms that we will need to build
the gem on, since we still have to rely on puppetlabs/packaging to ship
platform-specific versions of the gem.

Note you can pass in a value for the environment variable GEM_PLATFORM
(e.g. x86-mingw32) to build the gem for a specific platform that's
different than the build machine.

Finally, this PR also removes the "files" entry from
ext/project_data.yaml, because puppetlabs/packaging now has the ability
to TAR up the entire working directory by default, if a "files" entry is
not specified (which is what we want to do, anyways). This way,
the .gemspec is self-contained since we can now specify the files we
need there instead.