(PUP-8693) Add a compare(a,b) function to puppet
This adds the function compare to puppet. It returns -1, 0, or 1 if a is
before b, same as b, or after b. The function accepts the comparable
types Numeric, String, Semver, Timestamp and Timespan. A third argument
is accepted when both values are strings to control if the compare
ignores case or not.

This also adds cases to the ModelLabelProvider for missing data types.