(PUP-8603) Add the max() function from stdlib to puppet
This adds a stdlib compatible max() function to puppet. It deprecates
behaviour in the stdlib version that:

* converts strings to number to perform numeric compare if possible
* converts any value to string before comparing (possibly numeric)

A new feature is added that allows a lambda to be give to take over the
comparison of values - this gives users the ability to use sane
comparison instead of the deprecated stdlib behavior.