(PUP-8696) Explicitly pass current environment to autoloader
Update the autoloader methods to accept an optional environment parameter
defaulting to the current environment. Update call sites to pass the current
environment explicitly.

The call sites are found by grepping and also temporarily making the environment
argument required to see which code/tests fail.

The environment parameter was made optional instead of mandatory, because there
are large number of 3x functions in modules that call `loader.loadall`, as
apparently that was necessary at one point, but that was removed from
built-in functions in PUP-4058.