(PUP-8696) Use configured environment to load the app
Applications can be distributed in modules, so we need to use the
configured environment's modulepath to find them. If the configured
environment doesn't exist, then we need to fall back to the
current environment, which is *root* at this point in the bootstrap
process. The root environment has an empty modulepath, so we'll only be
able to find applications that are in puppet.