(PUP-8603) Make max function handle single array as if it was splatted
Before this it was required to splat an array as arguments to max
in order to find the max value. Since it is an otherwise meaningless
operation to call max with a single array (it would return the array)
that special case is now taken as a splat of the array's content.

Thus making `max([1,2,3])` equivalent to `max(*[1,2,3])`