(PUP-8603) Make min function handle single array as if it was splatted
Before this it was required to splat an array as arguments to min
in order to find the min value. Since it is an otherwise meaningless
operation to call min with a single array (it would return the array)
that special case is now taken as a splat of the array's content.

Thus making `min([1,2,3])` equivalent to `min(*[1,2,3])`