(PUP-8655) Remove host parameter for validator initialization
Previous to this commit, the DefaultValidator took a parameter of a
Puppet::SSL::Host to initialize. This commit removes that requirement and does
the lookup for the ssl_host if #setup_connection method is called. This allows
for any arbitrary class to still use this class for for SSL verification without
requiring the need for a Puppet::SSL::Host.