(PUP-8604) Add the downcase() function from stdlib to puppet
This adds the function `downcase()` from stdlib to puppet.
It is a re-implementation using the 4.x function API.

There is one difference in that the stdlib function would return an
empty string as the result of `downcase(undef)`, where the new version
instead errors. The stdlib version would however error if given an Array
or Hash with `undef` values as only top level arguments have `undef`
replaced with empty string in the 3.x function API.

This implementation adds being able to downcase a Hash (as that is
consistent with how the stdlib `upcase` function works - but was missing
in the stdlib implementation).