(PUP-8603) Fix bad regexp in min() max() for detecting float
The regexp (copied from stdlib version of these functions) had an error
in the regexp for detected a decimal point as there was no escape and as
a consequence any char would match.

Thus, some unintended valid float strings where matched - those that had
an e or E for scientific notation, or those that used a _ as a separator
for arbitrary digit grouping. All other characters would lead to an
error as the false positive would lead to creation of a float from a
string that could not possibly be a float.

With the fix, such strings would now be compared as strings instead of
error.