(PUP-8655) Update usage of peer_certs
In PUP-8655 we updated the DefaultValidator to directly use
OpenSSL::X509::Certificate objects in its peer_certs.

The Puppet::Network::HTTP::Connection object was using this
variable directly when constructing an error message, expecting
a Puppet::SSL::Base subclass.

This updates the Connection to expect the
OpenSSL::X509::Certificate objects when constructing its error message,
and to use the Puppet::SSL::Certificate class as a singleton for the
subject_alt_names_for() helper.