(PUP-8488) Allow task implementations to have any name
Previously, task implementations were required to have basenames
matching the task name. That meant it was impossible to have multiple
implementations of the same task with the same file extension. It also
prevented a single executable from being used by multiple tasks.

Loading a task with both metadata and implementations was previously
accomplished in the past by defining a "match_many" behavior for tasks,
indicating that the loader should look for multiple files matching the
task name, not just one.

We now allow the task implementations list to refer to any file in the
tasks/ directory of the module, as long as it exists. This is
accomplished by redefining the "match_many" behavior to instead be a
"fuzzy_match", in which case the loader will find all files that *might*
be part of the task (which in this case is every file in the tasks
directory). Reading the metadata and determining which files are
*actually* part of the task is delegated to the TaskInstantiator.