(PUP-8745) Make match() handle type aliases and Variant type
Before this, the match function would error if given a type alias or a
variant type. This is bad because it means that regexp based types
cannot be used to match with a result other than true/false.

Now, Variant and aliased types are resolved and work as expected when
they resolve to a regexp based type.