(PUP-8705) Add SRV record support with caching
This commit adds SRV support to Puppet::Rest::Client. It also updates
our existing SRV support to cache the DNS record list, to cut down on
the number of DNS lookups we do. Note that to preserve previous
behavior, indirector-based requests will still perform DNS resolution
every request.

The Network::Resolver is now a class that stores the result of the DNS
lookup. TTL checking for this cache will come in a later commit.

The Rest::Client will cache the first server it can talk to during
server resolution and only revisit the DNS listings if that server
can no longer be reached.