(maint) Add an evaluate method to Compiler spec helper
Before this it was required to use compile_to_catalog and
check for resources with a given title. Now for tests where
only a simple evaluation result is wanted it is possible to
instead call `evaluate` - which takes two sources - a set up
`code` and `source`. If only `code` is present, it returns the
the catalog, and if `source` is present if first compiles the
catalog and then evaluates the `source` and returns the
result of that evaluation.
The `evaluate` method also accepts a `variables` hash that
sets top scope variables before the compilation takes place.