(PUP-7822) Add a get() function to puppet
Before this it was somewhat complicated to use the same dot
navigation syntax as used elsewhere to dig out values as
the `dig()` function requires an array of values for digging.

The new `get()` function can get from a variable name, or
from a value, supports a default value, and can handle the typical
errors of attempt to navigate into array with string key, and
if navigating into something that is not a Collection.