(PUP-7822) Make dig() raise error with pcore Error datatype
In order for a caller to get the details from errors in dig
it needs to return an error with details such as `walked_path`.
For that it is most suitable to create an instance of the pcore
rich data type Error as it can be directly used in the puppet language.

A new Puppet::Error subtype ErrorWithData was added for this purpose.
The dig() function now raises that error instead of just
ArgumentError. The dig() function now also raises a specific error when
trying to use a string index with an Array.

The get function now passes the instance of Error on to a given block
instead of calling with issues code strings.