(PUP-1952) Fix schedule type with jruby
Due to https://github.com/jruby/jruby/issues/1795 this was failing to
calculate overlaps correctly on jruby. We shouldn't be doing our own
date math like this anyway so I rewrote the `match?` function. This
incidentally fixes a bug with arrays of ranges if an early range spanned
midnight but did not match and a later range did, the check on the first
range would return out of the function with false and the second range
would never be checked.