(PUP-8696) Use current environment during autoloading
Use the current environment instead of the configured environment when
searching modules. Often these are the same, except for server and agent
applications.

The server should autoload types in the context of the agent's environment for
whom the server is compiling a catalog. Note the server only autoloads
the type if the pcore representation hasn't been generated.

The agent should never autoload types from locally installed modules,
instead always preferring pluginsync'ed types & providers, which are
consistent with the downloaded catalog. This is achieved, because the
agent's current environment is an instance of
Puppet::Node::Environment::Remote with an empty modulepath. So locally
installed modules are ignored, and the libdir is used instead.