(PUP-4242) Correctly validate named definitions in plans
This bit of code validates that the name of a definition in a Puppet
manifest conforms to the naming convention of
<module>::<dir>::<filename>. However, it determines the expected module
name by searching up the path until it finds a known directory name that
might contain a .pp file. Previously, it was only looking for
'manifests', 'functions', and 'types'. This meant that plans couldn't be
loaded, since it couldn't find the expected module name.