(PUP-8696) Deprecate Puppet::Parser::Functions.autoloader
3x functions sometimes call the Puppet::Parser::Functions.autoloader directly to
ensure a dependent function is loaded before trying to call it. To maintain
backwards compatibility with existing 3x functions in modules, this adds a
AutoloaderDelegate to issue a deprecation warning (unless strict is off). The
delegate preserves the autoloader's API for loadall, load and loaded?. The first
2 take an optional environment parameter, while the last doesn't.

The Runtime3Support.call_function method calls Functions.function, so call the
delegatee to avoid the deprecation warning.