(PUP-7141) Put back dup of frozen when not in master mode
When puppet is running as a master it does not do apply of any unsafe
resources (resource types that mutate parameters). This commit brings
back "copy if string is frozen" if puppet is not running in master
mode. Thus making taking the safe and optimized route on compile masters
while allowing unsafe code to run on the agents unaltered).